backend/pyproject.toml
[tool.poetry]
name = "daiichi-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
langchain = "^0.3.26"
langchain-openai = "^0.3.27"
langchain-community = "^0.3.26"
azure-search-documents = "^11.5.3"
azure-identity = "^1.23.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
dotenv = "^0.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"
black = "^25.1.0"
ruff = "^0.12.2"

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"